Rename service-resource to resource-link
This better matches the language used in Fastly's web UI.
