+++ pkg/commands/resourcelink/create.go
@@ -1,4 +1,4 @@
-package serviceresource
+package resourcelink
 
 import (
 	"io"
