+++ pkg/commands/resourcelink/doc.go
@@ -1,4 +1,4 @@
-// Package serviceresource contains commands to inspect and
+// Package resourcelink contains commands to inspect and
 // manipulate service resource links.
 // https://developer.fastly.com/reference/api/services/resource/
-package serviceresource
+package resourcelink
